Styled the pages.
