+++ components/AppHeader.vue
@@ -93,11 +93,11 @@ const mobileMenuOpen = ref(false);
         <div class="flex items-center justify-between">
           <a href="#" class="-m-1.5 p-1.5">
             <span class="sr-only">Bach.Software</span>
-            <img
+            <!-- <img
               class="h-8 w-auto"
               src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
               alt=""
-            />
+            /> -->
           </a>
           <button
             type="button"
