+++ content/posts/1.md
@@ -2,13 +2,14 @@
 slug: "refReactiveDefault-title"
 title: "refReactiveDefault title"
 description: "meta description of the page"
-readTime: 10 min
+category: Vue.js
 author: jeroenbach
-date: Aug 27, 2024
-imgCover: first-blog-post.jpg
+date: 2024-08-27
+imgCoverUrl: /posts/1/cover.jpeg
+readTime: 10 min
 ---
 
-A helper tool that will let you stick to true component design.
+A helper tool that will let you stick to true component design. Implement this component alsdfksa ldfjlksadlfjlaskdf lsa ldajklfds
 
 <!--more-->
