+++ package.json
@@ -21,6 +21,7 @@
     "@heroicons/vue": "^2.1.5",
     "@vueuse/core": "^11.0.0",
     "animate.css": "^4.1.1",
+    "dayjs": "^1.11.13",
     "pinia": "^2.2.2",
     "sitemap": "^8.0.0",
     "typescript": "^5.5.4"
