+++ pages/posts/index.vue
@@ -1,71 +1,9 @@
 <script lang="ts" setup>
 import PostsContext from "~/contexts/PostsContext.vue";
 import AuthorsContext from "~/contexts/AuthorsContext.vue";
-
-// const { data: posts } = await useAsyncData("posts", async () => {
-//   const posts = await queryContent("posts").find();
-//   return posts;
-// });
-// const search = ref("");
-// const results = await searchContent(search);
-
-// const posts = [
-// {
-//   id: 1,
-//   title: "Boost your conversion rate",
-//   href: "/posts/1-boost-your-conversion-rate",
-//   description:
-//     "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
-//   date: "Mar 16, 2020",
-//   datetime: "2020-03-16",
-//   category: { title: "Marketing", href: "#" },
-//   author: {
-//     name: "Michael Foster",
-//     role: "Co-Founder / CTO",
-//     href: "#",
-//     imageUrl:
-//       "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-//   },
-// },
-// {
-//   id: 2,
-//   title: "Boost your conversion rate",
-//   href: "/posts/1-boost-your-conversion-rate-123",
-//   description:
-//     "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
-//   date: "Mar 16, 2020",
-//   datetime: "2020-03-16",
-//   category: { title: "Marketing", href: "#" },
-//   author: {
-//     name: "Michael Foster",
-//     role: "Co-Founder / CTO",
-//     href: "#",
-//     imageUrl:
-//       "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-//   },
-// },
-// {
-//   id: 3,
-//   title: "Boost your conversion rate",
-//   href: "/posts/1-boost-your-conversion-rate-123",
-//   description:
-//     "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
-//   date: "Mar 16, 2020",
-//   datetime: "2020-03-16",
-//   category: { title: "Marketing", href: "#" },
-//   author: {
-//     name: "Michael Foster",
-//     role: "Co-Founder / CTO",
-//     href: "#",
-//     imageUrl:
-//       "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-//   },
-// },
-// More posts...
-// ];
 </script>
 <template>
-  <div class="bg-white py-6 sm:py-12">
+  <div class="py-6 sm:py-12">
     <div class="mx-auto max-w-7xl px-6 lg:px-8">
       <div class="mx-auto max-w-2xl lg:mx-0">
         <h2 class="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
@@ -82,52 +20,67 @@ import AuthorsContext from "~/contexts/AuthorsContext.vue";
           <template #post="{ post }">
             <AuthorsContext :userName="post.author">
               <template #author="{ author }">
-                <article
-                  class="flex max-w-xl flex-col items-start justify-between"
-                >
-                  <div class="flex items-center gap-x-4 text-xs">
-                    <time :datetime="post.date" class="text-gray-500">{{
-                      post.date
-                    }}</time>
-                    <NuxtLink
-                      :to="'post.category.href'"
-                      class="relative z-10 rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
-                      >{{ "post.category.title" }}</NuxtLink
-                    >
-                  </div>
-                  <div class="group relative">
-                    <h3
-                      class="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600"
-                    >
-                      <NuxtLink :to="post.url">
-                        <span class="absolute inset-0" />
-                        {{ post.title }}
-                      </NuxtLink>
-                    </h3>
-                    <ContentRenderer
-                      :value="post"
-                      :excerpt="true"
-                      class="mt-5 line-clamp-3 text-sm leading-6 text-gray-600"
-                    />
-                  </div>
-                  <div class="relative mt-8 flex items-center gap-x-4">
+                <article class="flex flex-col items-start gap-8 lg:flex-row">
+                  <div class="w-full lg:w-64">
                     <NuxtPicture
-                      :src="author.imageUrl"
-                      sizes="50px"
-                      :alt="author.fullName"
-                      :imgAttrs="{ class: 'h-10 w-10 rounded-full bg-gray-50' }"
+                      :src="post.imgCoverUrl"
+                      :alt="post.title"
+                      sizes="sm:640px, lg:256px"
+                      :imgAttrs="{
+                        class:
+                          'aspect-16/9 sm:aspect-2/1 lg:aspect-1/1 w-full rounded-2xl bg-slate-200 object-cover',
+                      }"
                     />
-                    <div class="text-sm leading-6">
-                      <p class="font-semibold text-gray-900">
-                        <NuxtLink
-                          class="text-sky-600 hover:text-sky-800 dark:text-sky-500 dark:hover:text-sky-300"
-                          :to="author.linkedin"
-                        >
+                  </div>
+                  <div
+                    class="flex max-w-xl flex-col items-start justify-between"
+                  >
+                    <div class="flex items-center gap-x-4 text-xs">
+                      <time :datetime="post.date" class="text-gray-500"
+                        >{{ formatDate(post.date) }}
+                      </time>
+                      <NuxtLink
+                        :to="`/posts?category=${post.category}`"
+                        class="relative z-10 rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
+                        >{{ post.category }}</NuxtLink
+                      >
+                    </div>
+                    <div class="group relative max-w-xl">
+                      <h3
+                        class="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600"
+                      >
+                        <NuxtLink :to="post.url">
                           <span class="absolute inset-0" />
-                          {{ author.fullName }}
+                          {{ post.title }}
                         </NuxtLink>
-                      </p>
-                      <p class="text-gray-600">{{ author.role }}</p>
+                      </h3>
+                      <ContentRenderer
+                        :value="post"
+                        :excerpt="true"
+                        class="mt-5 line-clamp-3 text-sm leading-6 text-gray-600"
+                      />
+                    </div>
+                    <div class="relative mt-8 flex items-center gap-x-4">
+                      <NuxtPicture
+                        :src="author.imageUrl"
+                        sizes="50px"
+                        :alt="author.fullName"
+                        :imgAttrs="{
+                          class: 'h-10 w-10 rounded-full bg-gray-50',
+                        }"
+                      />
+                      <div class="text-sm leading-6">
+                        <p class="font-semibold text-gray-900">
+                          <NuxtLink
+                            class="text-sky-600 hover:text-sky-800 dark:text-sky-500 dark:hover:text-sky-300"
+                            :to="author.linkedin"
+                          >
+                            <span class="absolute inset-0" />
+                            {{ author.fullName }}
+                          </NuxtLink>
+                        </p>
+                        <p class="text-gray-600">{{ author.role }}</p>
+                      </div>
                     </div>
                   </div>
                 </article>
