+++ tailwind.config.js
@@ -33,6 +33,11 @@ export default {
           },
         }
       }),
+      aspectRatio: {
+        '16/9': '16 / 9',
+        '2/1': '2 / 1',
+        '1/1': '1 / 1',
+      }
     }, 
   },
   plugins: [
