+++ types/Post.ts
@@ -2,12 +2,14 @@ import type { MarkdownParsedContent } from "@nuxt/content";
 
 export interface PostSummary
   extends Pick<MarkdownParsedContent, "_path" | "excerpt"> {
-  title?: string;
   slug?: string;
-  readTime?: string;
+  title?: string;
+  description: string;
+  category?: string;
   author?: string;
   date?: string;
-  imgCover?: string;
+  imgCoverUrl?: string;
+  readTime?: string;
 }
 
 export interface Post extends PostSummary, MarkdownParsedContent {}
