+++ utils/checks.ts
@@ -1,0 +1,2 @@
+export const isFalseOrUndefined = (value: boolean | null | undefined) =>
+  value === false || value === undefined;
